+++ Source/Data/ChangelogData.lua
@@ -7,6 +7,7 @@ ChangelogData = {
         version = "0.3.0",
         changes = {
             { category = "Added", text = "Dungeon Quest filters." },
+            { category = "Added", text = "Slash Command support (/fdb)" },
             { category = "Changed", text = "Quest Effort is now Low, Medium or High." },
             { category = "Changed", text = "Dungeon Quest now shows Dungeon Level range." },
             { category = "Changed", text = "Quest Effort is now Low, Medium or High." },